+++ CHANGELOG.md
@@ -1,11 +1,10 @@
 # Change Log
 
+- handling conversion -9 error
 - nodejs: wopi formsubmit icon
-- php: handling conversion -9 error
 - nodejs: tabs menu
 - nodejs: delete all files
 - change insert image
-- nodejs: handling conversion -9 error
 - different goback for users
 - nodejs: converting function on index page
 - nodejs: close editor